+++ android/app/src/main/kotlin/com/example/realm_roller/MainActivity.kt
@@ -2,5 +2,4 @@ package com.itaybenhaim.realm_roller
 
 import io.flutter.embedding.android.FlutterActivity
 
-class MainActivity: FlutterActivity() {
-}
+class MainActivity: FlutterActivity()