+++ lib/pages/npcs/npc_view/npc_view.dart
@@ -3,6 +3,7 @@ import 'package:randpg/entities/npcs.dart';
 import 'package:randpg/string_manipulations.dart';
 import 'package:realm_roller/custom_widgets/entity_pages/entity_page.dart';
 import 'package:realm_roller/custom_widgets/expanded_paragraphs/expanded_paragraph.dart';
+import 'package:realm_roller/custom_widgets/tiles/companion/companion_tile.dart';
 
 import '../../../assets_handlers/image_path_finders.dart';
 import 'npc_description_formatters.dart';
@@ -43,7 +44,6 @@ class NpcView extends StatelessWidget {
                     .toList(),
               ),
             ),
-            const SizedBox(height: 24),
             ExpandedParagraph(
               title: "Personality",
               icon: Icons.psychology_alt,
@@ -54,6 +54,24 @@ class NpcView extends StatelessWidget {
               ),
             ),
             const SizedBox(height: 24),
+            if (npc.companions.isNotEmpty)
+              ExpandedParagraph(
+                title: "Companion${npc.companions.length == 1 ? "" : "s"}",
+                icon: Icons.pets,
+                child: Column(
+                  children: npc.companions
+                      .map(
+                        (e) => Column(
+                          children: [
+                            CompanionTile(companion: e),
+                            const SizedBox(height: 16),
+                          ],
+                        ),
+                      )
+                      .toList(),
+                ),
+              ),
+            if (npc.companions.isNotEmpty) const SizedBox(height: 24),
           ],
         ),
       ),