+++ lib/theme/theme_data.dart
@@ -1,19 +1,22 @@
 import 'package:flutter/material.dart';
 import 'package:realm_roller/theme/text_theme.dart';
 
+const _swatch = MaterialColor(0xFFFF2B0B, {
+  50: Color(0xFFFFEAEC),
+  100: Color(0xFFFFCACC),
+  200: Color(0xFFFA948E),
+  300: Color(0xFFF26861),
+  400: Color(0xFFFB4236),
+  500: Color(0xFFFF2B0B),
+  600: Color(0xFFF11A11),
+  700: Color(0xFFDF000B),
+  800: Color(0xFFD20001),
+  900: Color(0xFFC50000),
+});
+
 final themeData = ThemeData(
-  primarySwatch: const MaterialColor(0xFFFF2B0B, {
-    50: Color(0xFFFFEAEC),
-    100: Color(0xFFFFCACC),
-    200: Color(0xFFFA948E),
-    300: Color(0xFFF26861),
-    400: Color(0xFFFB4236),
-    500: Color(0xFFFF2B0B),
-    600: Color(0xFFF11A11),
-    700: Color(0xFFDF000B),
-    800: Color(0xFFD20001),
-    900: Color(0xFFC50000),
-  }),
+  primarySwatch: _swatch,
+  primaryColor: _swatch.shade500,
   textTheme: textTheme,
   primaryTextTheme: textTheme,
 );